+++ CHANGES.rst
@@ -114,6 +114,8 @@ Specviz
 Specviz2d
 ^^^^^^^^^
 
+- Fix subset linking/displaying between pixel/wavelength in Specviz2d viewers. [#2736]
+
 4.1.1 (2025-01-31)
 ==================
 
@@ -747,6 +749,8 @@ Specviz
   display label for the y axis in spectral viewer. Previously it was hard coded
   to always display ``flux density`` no matter the input unit. [#2703]
 
+Specviz2d
+^^^^^^^^^
 
 3.8.2 (2024-02-23)
 ==================